Update coverage report and line hit tracking
